Use an edit file to share more common code with the main Paragraph web code.
